Add WebHook Service
This service has the main scope of to manage the webhook of GitHub and Gitab
GitLab. I'ts a HTTP server so there are two path '/gitlab' and '/github'.
When there is request , the response it's the same , status code 200.
Because it's not important the response for GitHub or GitLab.
The most important thing is the body of HTTP POST request that receive.
Because the service get a JSON with many and many information about 'push event' and so we need to parse this JSON.
The JSON is an object so it has many keys and we use only the key 'ref' to get the branch that trigger the push event.
Now that we have the branch we can make a switch, or if you want an if-else block.
There are too many cases like how many branches do you have.
Inside the case you can update the repo with the command 'exec' tha give you the possibility to run bash script to update the repo, to install new depencies, to restart the service of systemctl that run the server-web, tasks.